backend/requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
werkzeug==2.3.7
PyJWT==2.8.0
